+++ contracts/orbital-core/src/contract.rs
@@ -1,36 +1,28 @@
 use crate::{
     admin_logic::admin,
-    error::ContractError,
     icq::{self},
     state::{ClearingAccountConfig, OrbitalDomainConfig, UserConfig, USER_NONCE},
     user_logic::user,
-    utils::{
-        extract_ica_identifier_from_port, fees::flatten_ibc_fees_amt, get_ica_identifier,
-        OpenAckVersion,
-    },
+    utils::{extract_ica_identifier_from_port, get_ica_identifier, OpenAckVersion},
 };
 use crate::{
     msg::{ExecuteMsg, InstantiateMsg, MigrateMsg, QueryMsg},
     state::{CLEARING_ACCOUNTS, ORBITAL_DOMAINS, USER_CONFIGS},
 };
-use cosmos_sdk_proto::cosmos::base::v1beta1::Coin as ProtoCoin;
 #[cfg(not(feature = "library"))]
 use cosmwasm_std::entry_point;
 use cosmwasm_std::{
-    ensure, to_json_binary, Addr, Binary, Coin, Deps, DepsMut, Env, MessageInfo, Reply, Response,
-    StdError, StdResult, Uint64,
+    to_json_binary, Addr, Binary, Deps, DepsMut, Env, MessageInfo, Reply, Response, StdError,
+    StdResult, Uint64,
 };
 use cw2::set_contract_version;
 use cw_ownable::{get_ownership, initialize_owner};
-use cw_utils::must_pay;
 use neutron_sdk::{
-    bindings::{msg::NeutronMsg, query::NeutronQuery, types::ProtobufAny},
+    bindings::{msg::NeutronMsg, query::NeutronQuery},
     interchain_queries::v047::queries::{query_balance, BalanceResponse},
-    query::min_ibc_fee::query_min_ibc_fee,
     sudo::msg::SudoMsg,
     NeutronResult,
 };
-use prost::Message;
 
 pub const CONTRACT_NAME: &str = "orbital-core";
 pub const CONTRACT_VERSION: &str = env!("CARGO_PKG_VERSION");
@@ -80,77 +72,9 @@ pub fn execute(
             denoms,
         } => icq::register_balances_query(connection_id, addr, denoms, update_period),
         ExecuteMsg::UserWithdrawFunds { domain, coin, dest } => {
-            try_withdraw_from_remote_domain(deps, info, domain, coin, dest)
+            user::try_withdraw_from_remote_domain(deps, info, domain, coin, dest)
         }
     }
-}
-
-fn try_withdraw_from_remote_domain(
-    deps: ExecuteDeps,
-    info: MessageInfo,
-    domain: String,
-    coin: Coin,
-    dest: String,
-) -> NeutronResult<Response<NeutronMsg>> {
-    let user_config = USER_CONFIGS.load(deps.storage, info.sender.to_string())?;
-
-    // first we validate that user is registered to the domain from which they
-    // want to withdraw funds from
-    ensure!(
-        user_config.registered_domains.contains(&domain),
-        ContractError::UserNotRegisteredToDomain(domain)
-    );
-
-    // validate that IBC fees are covered by the caller
-    let min_ibc_fee = query_min_ibc_fee(deps.as_ref())?;
-    let total_fee_amt = flatten_ibc_fees_amt(&min_ibc_fee.min_fee);
-    let paid_amt = must_pay(&info, "untrn").map_err(ContractError::FeePaymentError)?;
-
-    ensure!(
-        paid_amt >= total_fee_amt,
-        ContractError::Std(StdError::generic_err("insufficient fee coverage"))
-    );
-
-    // derive the port associated with user's clearing account
-    let ica_identifier = get_ica_identifier(user_config.id, domain.to_string());
-
-    let user_clearing_acc_config = CLEARING_ACCOUNTS
-        .load(deps.storage, ica_identifier.to_string())?
-        .ok_or_else(|| ContractError::UserNotRegisteredToDomain(domain))?;
-
-    // generate the transfer message to be executed on target domain
-    let proto_coin = ProtoCoin {
-        denom: coin.denom,
-        amount: coin.amount.to_string(),
-    };
-    let bank_msg = cosmos_sdk_proto::cosmos::bank::v1beta1::MsgSend {
-        from_address: user_clearing_acc_config.addr,
-        to_address: dest,
-        amount: vec![proto_coin],
-    };
-
-    let proto_msg = generate_proto_msg(bank_msg, "/cosmos.bank.v1beta1.MsgSend")?;
-
-    let withdraw_tx: NeutronMsg = NeutronMsg::submit_tx(
-        user_clearing_acc_config.controller_connection_id,
-        ica_identifier,
-        vec![proto_msg],
-        "".to_string(),
-        60,
-        min_ibc_fee.min_fee,
-    );
-
-    Ok(Response::default().add_message(withdraw_tx))
-}
-
-fn generate_proto_msg(msg: impl Message, type_url: &str) -> NeutronResult<ProtobufAny> {
-    let buf = msg.encode_to_vec();
-
-    let any_msg = ProtobufAny {
-        type_url: type_url.to_string(),
-        value: Binary::from(buf),
-    };
-    Ok(any_msg)
 }
 
 #[entry_point]
